refact: rename ui package
