+++ app/src/main/java/com/github/felipecastilhos/pokedexandroid/features/pokemon/ui/components/AboutPanel.kt
@@ -1,8 +1,7 @@
-package com.github.felipecastilhos.pokedexandroid.features.pokemon.presentation.components
+package com.github.felipecastilhos.pokedexandroid.features.pokemon.ui.components
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.*
-import androidx.compose.material.Divider
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
@@ -10,13 +9,10 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.ColorFilter
 import androidx.compose.ui.graphics.painter.Painter
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.github.felipecastilhos.pokedexandroid.core.ui.components.DividerVertical
 import com.github.felipecastilhos.pokedexandroid.core.ui.theme.GrayscalePalette
-import com.github.felipecastilhos.pokedexandroid.core.ui.theme.PokedexAndroidTheme
 import com.github.felipecastilhos.pokedexandroid.core.ui.theme.PokedexTheme
