+++ app/src/main/java/com/github/felipecastilhos/pokedexandroid/showcase/components/AboutPanelShowcase.kt
@@ -1,13 +1,13 @@
-package com.github.felipecastilhos.pokedexandroid.features.pokemon.presentation.uicatalog.components
+package com.github.felipecastilhos.pokedexandroid.features.pokemon.ui.uicatalog.components
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import com.github.felipecastilhos.pokedexandroid.core.ui.theme.PokedexAndroidTheme
 import com.github.felipecastilhos.pokedexandroid.core.ui.theme.PokedexTheme
-import com.github.felipecastilhos.pokedexandroid.features.pokemon.presentation.components.AboutPanelIconInfo
-import com.github.felipecastilhos.pokedexandroid.features.pokemon.presentation.components.AboutPanelLayout
-import com.github.felipecastilhos.pokedexandroid.features.pokemon.presentation.components.AboutPanelListInfo
+import com.github.felipecastilhos.pokedexandroid.features.pokemon.ui.components.AboutPanelIconInfo
+import com.github.felipecastilhos.pokedexandroid.features.pokemon.ui.components.AboutPanelLayout
+import com.github.felipecastilhos.pokedexandroid.features.pokemon.ui.components.AboutPanelListInfo
 
 @Preview(name = "About panel", group = "About Panel")
 @Composable
